fix: 뉴스 데이터 수정 및 코드 리팩토링
