+++ src/_hooks/fetcher/news/useGetNewsDataList.ts
@@ -1,31 +1,13 @@
-// hooks/useGetNewsDataList.ts
 import { useQuery, UseQueryResult } from "@tanstack/react-query";
 import fetchNewsDataList from "@/services/news/fetchNewsDataList";
 import { NewsItemType } from "@/app/(route)/news/_types/newsItemType";
-
-export interface NewsQueryParams {
-  page?: string;
-  size?: number;
-  timePeriod?: "DAILY" | "WEEKLY" | "MONTHLY" | "YEARLY";
-  category?: "" | "FOOTBALL" | "BASEBALL" | "ESPORTS";
-  orderType?: "DATE" | "COMMENT" | "VIEW";
-  content?: string;
-  withPageInfo?: boolean;
-}
-
-export interface NewsListWithPageInfo {
-  content: NewsItemType[];
-  pageInfo: {
-    currentPage: number;
-    totalPage: number;
-    totalElements: number;
-  };
-}
+import { NewsDataProps } from "@/app/(route)/main/_types/NewsDataProps";
+import { NewsListWithPageInfo } from "@/app/(route)/main/_types/NewsDataProps";
 
 type NewsDataReturnType = NewsListWithPageInfo | NewsItemType[];
 
 const useGetNewsDataList = (
-  params: NewsQueryParams = {}
+  params: NewsDataProps = {}
 ): UseQueryResult<NewsDataReturnType> => {
   return useQuery<NewsDataReturnType>({
     queryKey: ["newsDataList", params],
